[FEAT] 잔액 서비스 구현 + 단위테스트 추가

+++ src/main/java/kr/hhplus/be/server/controller/balance/dto/BalanceResponseDTO.java
@@ -5,15 +5,17 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class BalanceResponseDTO {
     private long userId;
-    private int currentBalance;
+    private BigDecimal currentBalance;
 
     @Builder
-    public BalanceResponseDTO(long userId, int currentBalance) {
+    public BalanceResponseDTO(long userId, BigDecimal currentBalance) {
         this.userId = userId;
         this.currentBalance = currentBalance;
     }
